feature: Get single character information
After retrieving a list of star wars character effort was also made to
get information about each one of them. This has been achieved with this
changes. Also, a landing page was added so that what users get to see
isn't the characters when they first open the app.
Now as it stands, only the first page of the characters is being retrieved
 from the api, effor will be put into retrieving all the availale characters
 by going through all the pages provided by the api.